wasd

+++ GenetiskAlgoritme/src/main/java/org/openjfx/GenetiskAlgoritme/Items.java
@@ -13,7 +13,7 @@ public class Items {
 
     public float price;
 
-    public float weight;
+    public int weight;
 
     private int type;
